# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Gerenciamento de Hospedagem",
        "description": "Documentação do Gerenciamento de Hospedagem",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "luizcsdev@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "OpenApi host"
        }
    ],
    "paths": {
...
        "/api/v1/type-rooms": {
            "get": {
                "tags": [
                    "Tipo de Quarto"
                ],
                "summary": "Listar Tipos de Quarto",
                "operationId": "75e93c04f5b6ac9e3d49b0d4bdc07e7c",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token JWT gerado durante a autenticação. Deve ser enviado no formato: Bearer {token}",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer jwt_token_aqui"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exibe uma lista paginada de Tipos de Quarto, com a possibilidade de carregar atributos relacionados e aplicar filtros."
                    },
                    "401": {
                        "description": "Token ausente ou inválido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "erro": {
                                            "type": "string",
                                            "example": "Token inválido ou ausente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Missing Data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tipo de Quarto"
                ],
                "summary": "Criar um tipo de quarto",
                "operationId": "b58057876eb3b5343a6d9d1057d4f834",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token JWT gerado durante a autenticação. Deve ser enviado no formato: Bearer {token}",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer jwt_token_aqui"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dados para criação de um tipo de quarto",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "O nome do tipo de quarto",
                                        "type": "string",
                                        "example": "Standarth"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tipos de quarto criado com sucesso."
                    },
                    "401": {
                        "description": "Token ausente ou inválido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "erro": {
                                            "type": "string",
                                            "example": "Token inválido ou ausente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Missing Data"
                    },
                    "500": {
                        "description": "Ocorreu um erro ao tentar recuperar os tipos de quarto.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "erro": {
                                            "type": "string",
                                            "example": "Ocorreu um erro ao tentar recuperar os tipos de quarto.."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/all-type-rooms": {
            "get": {
                "tags": [
                    "Tipo de Quarto"
                ],
                "summary": "Listar toods os tipos de quarto sem paginação",
                "operationId": "f2d14939c6390b1786a5c90eb033bc3e",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token JWT gerado durante a autenticação. Deve ser enviado no formato: Bearer {token}",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer jwt_token_aqui"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exibe uma lista com todos os tipos de quarto sem paginação."
                    },
                    "401": {
                        "description": "Token ausente ou inválido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "erro": {
                                            "type": "string",
                                            "example": "Token inválido ou ausente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nenhum tipo de quarto encontrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "erro": {
                                            "type": "string",
                                            "example": "Nenhum tipo de quarto encontrado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Missing Data"
                    },
                    "500": {
                        "description": "Ocorreu um erro ao tentar recuperar os tipos de quarto.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "erro": {
                                            "type": "string",
                                            "example": "Ocorreu um erro ao tentar recuperar os tipos de quarto.."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}